fix bugs

+++ logs/ticket_logger.py
@@ -16,14 +16,14 @@ async def clean_tickets(today):
                 if ticket['closed_at'] < today - timedelta(days=30):
                     filename = f"help_bot/tickets_log/ticket_{ticket['id']}/ticket_{ticket['id']}.json"
                     path = f"help_bot/tickets_log/ticket_{ticket['id']}"
-                    logger.info(ticket['closed_at'])
 
-                    try:
-                        os.remove(filename)  # Удаляет файл
-                        os.rmdir(path)  # Удаляет только пустую папку
-                        count += 1
-                    except Exception as err:
-                        logger.info(f"Ошибка удаления тикета №{ticket['id']}: {err}")
+                    if os.path.exists(filename) and os.path.exists(path):
+                        try:
+                            os.remove(filename)  # Удаляет файл
+                            os.rmdir(path)  # Удаляет только пустую папку
+                            count += 1
+                        except Exception as err:
+                            logger.info(f"Ошибка удаления тикета №{ticket['id']}: {err}")
 
             if count > 0:
                 logger.info(f"Удалены {count} тикетов")
